+++ banditpylib/learners/collaborative_learner/__init__.py
@@ -1,6 +1,7 @@
 from .utils import *
 from .lilucb_heur_collaborative import *
 from .collaborative_agent import *
+from .collaborative_master import *
 
 __all__ = ['CollaborativeLearner', 'CollaborativeAgent',
     'CollaborativeMaster']
